+++ model/model_appapi_group_content_object_item.go
@@ -13,7 +13,7 @@ package model
 type AppapiGroupContentObjectItem struct {
 	Content   string `json:"content,omitempty"`
 	Publisher string `json:"publisher,omitempty"`
-	TimeStamp int32  `json:"timeStamp,omitempty"`
+	TimeStamp int64  `json:"timeStamp,omitempty"`
 	TrxId     string `json:"trxId,omitempty"`
 	TypeUrl   string `json:"typeUrl,omitempty"`
 }
